modularizate functions and files
