+++ NNs/Optimizer/__init__.py
@@ -1,0 +1,2 @@
+from .Radam import RiemannianAdam
+from .rsgd import RiemannianSGD
