+++ parameters.py
@@ -1,8 +1,11 @@
-V = 26
-L = 2000
-N = 10000
-S = "abcdefghijklmnopqrstuvwxyz"[:V]
-VOCABULARY = [i for i in S]
-URL = f"../data/distances/distance_{V}_{L}.csv"
-FULLY_EUCLIDEAN = False
-URL_EMBEDDING = f"../data/embeddings/embedding_{V}_{L}.csv"
+from utils.data_Params import *
+
+##########################
+#####    Model Params
+##########################
+
+URL_EMBEDDING = f"data/embeddings/{V}_{L}.csv"
+EPOCHS = 500
+BATCH_SIZE = 128
+LEARNING_RATE = 0.001
+NUM_FEATURES = 4
