+++ utils/__init__.py
@@ -1,0 +1,1 @@
+from .general import generate_data
