+++ utils/functions.py
@@ -1,12 +1,9 @@
 import pandas as pd
 import numpy as np
-from parameters import *
+from .data_Params import *
 
 # funcion of the real distance in a tree between two nodes
-def metric(
-    x: int,
-    y: int,
-) -> float:
+def metric(x: int, y: int) -> float:
     h = 0
     for i in range(min(len(x), len(y))):
         if x[i] == y[i]:
@@ -15,11 +12,11 @@ def metric(
             break
     value = len(x) + len(y) - 2 * h
     # print(value/1000)
-    return value / 1000
+    return value
 
 
 # function to generate a random word
-def generate_word(max_length=L, vocabuary=["a", "b", "c"]) -> str:
+def generate_word(max_length: int = L, vocabuary: list = ["a", "b", "c"]) -> str:
     s = ""
     # print("maxlength", max_length)
     number = np.random.randint(0, max_length + 1)
@@ -28,21 +25,21 @@ def generate_word(max_length=L, vocabuary=["a", "b", "c"]) -> str:
     return s
 
 
-def embedding1(x: int):
+def embedding1(x: int) -> np.array:
     l = 0
     for i in range(len(x)):
         l += (i + 1) * (S.find(x[i]) + 1)
     return np.array(l)
 
 
-def embedding2(x: int):
+def embedding2(x: int) -> np.array:
     l = 0
     for i in range(len(x)):
         l += S.find(x[i]) + 1
     return np.array(l)
 
 
-def embed(x: int):
+def embed(x: int) -> np.array:
     l = np.zeros(L)
     for i in range(len(x)):
         l[i] = S.find(x[i])
